+++ packages/stable/indent-okta-webhook/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-canary.12](https://github.com/indentapis/integrations/compare/@indent/okta-webhook@0.0.1-canary.11...@indent/okta-webhook@0.0.1-canary.12) (2022-01-10)
+
+
+### Features
+
+* add okta-user ([2ddd300](https://github.com/indentapis/integrations/commit/2ddd300d509e003b57795e598ff945f5957d59f2))
+
+
+
+
+
+## [0.0.1-canary.11](https://github.com/indentapis/integrations/compare/@indent/okta-webhook@0.0.1-canary.10...@indent/okta-webhook@0.0.1-canary.11) (2022-01-10)
+
+
+### Features
+
+* add okta-user ([aaa0a6a](https://github.com/indentapis/integrations/commit/aaa0a6a33591fc0144b7e7e456cb88dc6115713b))
+
+
+
+
+
+## [0.0.1-canary.10](https://github.com/indentapis/integrations/compare/@indent/okta-webhook@0.0.1-canary.9...@indent/okta-webhook@0.0.1-canary.10) (2022-01-10)
+
+**Note:** Version bump only for package @indent/okta-webhook
+
+
+
+
+
+## [0.0.1-canary.9](https://github.com/indentapis/integrations/compare/@indent/okta-webhook@0.0.1-canary.8...@indent/okta-webhook@0.0.1-canary.9) (2022-01-10)
+
+
+### Bug Fixes
+
+* base handler ([9c3435b](https://github.com/indentapis/integrations/commit/9c3435b21e6ba13d27c9c1a7af85b98658202905))
+
+
+
+
+
+## [0.0.1-canary.8](https://github.com/indentapis/integrations/compare/@indent/okta-webhook@0.0.1-canary.7...@indent/okta-webhook@0.0.1-canary.8) (2022-01-10)
+
+**Note:** Version bump only for package @indent/okta-webhook
+
+
+
+
+
 ## [0.0.1-canary.7](https://github.com/indentapis/integrations/compare/@indent/okta-webhook@0.0.1-canary.6...@indent/okta-webhook@0.0.1-canary.7) (2022-01-04)